+++ actix-http/src/service.rs
@@ -580,6 +580,8 @@ where
                     peer_addr,
                 )),
             },
+
+            proto => unimplemented!("Unsupported HTTP version: {:?}.", proto)
         }
     }
 }